requirements_app.txt
streamlit==1.28.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0
python-dateutil==2.8.2
openai==0.28.1
anthropic==0.7.7
python-docx==0.8.11
Pillow==10.0.1
python-dotenv==1.0.0